# Avoid repeated memchr in SingleSearch