Avoid repeated memchr in SingleSearch

Repeated memchr in BMH slows down the literal search dramatically in
certain corner cases.