@bluss 2015-07-31
Evaluating benchmarks is a bit tricky. The benchmarks in the regex repository seem to be unchanged, or possibly slightly worse (may be fluctuations, benchmarking on a laptop).

I've collected some more pathological cases for substring search (no regular expressions), and my own simple B-M-H algorithm showed the same symptoms as yours.

Some of the examples speed up by up to 8-10 times with this change,

Here are benchmarks from both my [mess repo (exact commit linked)](https://github.com/bluss/scratchspace/tree/7458b8edc889a45bd7a75c2dddcc8a169a0882d0) and from the regex' own benchmarks.

```
Before

test aaabbb::regex_find                     ... bench:   3,227,704 ns/iter (+/- 223,193) = 92 MB/s
test allright::regex_find                   ... bench:   2,240,254 ns/iter (+/- 137,700) = 802 MB/s
test bad_naive::regex_find                  ... bench:       1,029 ns/iter (+/- 92) = 71 MB/s
test bad_naive_long::regex_find             ... bench:  32,255,378 ns/iter (+/- 1,607,109) = 3 MB/s
test bad_naive_long_reversed::regex_find    ... bench:      21,634 ns/iter (+/- 1,058) = 4622 MB/s
test bad_naive_reversed::regex_find         ... bench:          31 ns/iter (+/- 2) = 2387 MB/s
test bbbaaa::regex_find                     ... bench:   1,582,638 ns/iter (+/- 146,857) = 189 MB/s
test pathological_aa_100k::regex_find       ... bench:       3,834 ns/iter (+/- 460) = 26082 MB/s
test pathological_aab_100k::regex_find      ... bench:   3,286,591 ns/iter (+/- 195,221) = 91 MB/s
test pathological_two_way::regex_find       ... bench:       2,265 ns/iter (+/- 92) = 26490 MB/s
test pathological_two_way_rev::regex_find   ... bench:     325,389 ns/iter (+/- 27,369) = 184 MB/s
test short_long::regex_find                 ... bench:       2,810 ns/iter (+/- 134) = 907 MB/s
test short_short::regex_find                ... bench:          54 ns/iter (+/- 4) = 1037 MB/s


test bench::anchored_literal_long_match      ... bench:         356 ns/iter (+/- 37)
test bench::anchored_literal_long_non_match  ... bench:         154 ns/iter (+/- 7)
test bench::anchored_literal_short_match     ... bench:         295 ns/iter (+/- 21)
test bench::anchored_literal_short_non_match ... bench:         140 ns/iter (+/- 7)
test bench::easy0_1K                         ... bench:         443 ns/iter (+/- 113) = 2311 MB/s
test bench::easy0_1MB                        ... bench:     387,724 ns/iter (+/- 36,001) = 2704 MB/s
test bench::easy0_32                         ... bench:         118 ns/iter (+/- 15) = 271 MB/s
test bench::easy0_32K                        ... bench:      12,975 ns/iter (+/- 1,161) = 2525 MB/s
test bench::easy1_1K                         ... bench:         467 ns/iter (+/- 177) = 2192 MB/s
test bench::easy1_1MB                        ... bench:     467,720 ns/iter (+/- 31,924) = 2241 MB/s
test bench::easy1_32                         ... bench:         115 ns/iter (+/- 24) = 278 MB/s
test bench::easy1_32K                        ... bench:      12,945 ns/iter (+/- 2,609) = 2531 MB/s
test bench::hard_1K                          ... bench:      55,489 ns/iter (+/- 7,970) = 18 MB/s
test bench::hard_1MB                         ... bench: 118,775,246 ns/iter (+/- 21,646,321) = 8 MB/s
test bench::hard_32                          ... bench:       1,598 ns/iter (+/- 2,605) = 20 MB/s
test bench::hard_32K                         ... bench:   1,597,275 ns/iter (+/- 2,629,553) = 20 MB/s
test bench::literal                          ... bench:          27 ns/iter (+/- 49)
test bench::match_class                      ... bench:         121 ns/iter (+/- 180)
test bench::match_class_in_range             ... bench:         116 ns/iter (+/- 23)
test bench::match_class_unicode              ... bench:       2,953 ns/iter (+/- 6,624)
test bench::medium_1K                        ... bench:       3,241 ns/iter (+/- 5,722) = 315 MB/s
test bench::medium_1MB                       ... bench:   3,438,041 ns/iter (+/- 6,977,433) = 304 MB/s
test bench::medium_32                        ... bench:         182 ns/iter (+/- 8) = 175 MB/s
test bench::medium_32K                       ... bench:     110,612 ns/iter (+/- 232,915) = 296 MB/s
test bench::no_exponential                   ... bench:     579,349 ns/iter (+/- 873,794)
test bench::not_literal                      ... bench:       1,873 ns/iter (+/- 3,214)
test bench::one_pass_long_prefix             ... bench:         943 ns/iter (+/- 924)
test bench::one_pass_long_prefix_not         ... bench:         548 ns/iter (+/- 907)
test bench::one_pass_short_a                 ... bench:       3,216 ns/iter (+/- 3,733)
test bench::one_pass_short_a_not             ... bench:       2,041 ns/iter (+/- 3,424)
test bench::one_pass_short_b                 ... bench:       1,315 ns/iter (+/- 2,553)
test bench::one_pass_short_b_not             ... bench:       2,360 ns/iter (+/- 2,492)
test bench::replace_all                      ... bench:         356 ns/iter (+/- 511)

after, removed memchr:

running 13 tests
test aaabbb::regex_find                     ... bench:     332,135 ns/iter (+/- 19,754) = 903 MB/s
test allright::regex_find                   ... bench:   1,948,456 ns/iter (+/- 120,269) = 923 MB/s
test bad_naive::regex_find                  ... bench:         424 ns/iter (+/- 21) = 174 MB/s
test bad_naive_long::regex_find             ... bench:  38,245,484 ns/iter (+/- 1,295,389) = 2 MB/s
test bad_naive_long_reversed::regex_find    ... bench:      23,043 ns/iter (+/- 1,228) = 4339 MB/s
test bad_naive_reversed::regex_find         ... bench:          38 ns/iter (+/- 3) = 1947 MB/s
test bbbaaa::regex_find                     ... bench:     164,173 ns/iter (+/- 9,239) = 1827 MB/s
test pathological_aa_100k::regex_find       ... bench:       4,341 ns/iter (+/- 413) = 23036 MB/s
test pathological_aab_100k::regex_find      ... bench:     882,907 ns/iter (+/- 56,380) = 339 MB/s
test pathological_two_way::regex_find       ... bench:       2,545 ns/iter (+/- 124) = 23575 MB/s
test pathological_two_way_rev::regex_find   ... bench:     114,930 ns/iter (+/- 8,672) = 522 MB/s
test short_long::regex_find                 ... bench:       2,603 ns/iter (+/- 196) = 980 MB/s
test short_short::regex_find                ... bench:          57 ns/iter (+/- 5) = 982 MB/s



test bench::anchored_literal_long_match      ... bench:         273 ns/iter (+/- 58)
test bench::anchored_literal_long_non_match  ... bench:         136 ns/iter (+/- 24)
test bench::anchored_literal_short_match     ... bench:         275 ns/iter (+/- 85)
test bench::anchored_literal_short_non_match ... bench:         128 ns/iter (+/- 124)
test bench::easy0_1K                         ... bench:         576 ns/iter (+/- 93) = 1777 MB/s
test bench::easy0_1MB                        ... bench:     433,957 ns/iter (+/- 31,749) = 2415 MB/s
test bench::easy0_32                         ... bench:         106 ns/iter (+/- 8) = 301 MB/s
test bench::easy0_32K                        ... bench:      17,197 ns/iter (+/- 1,131) = 1905 MB/s
test bench::easy1_1K                         ... bench:         448 ns/iter (+/- 166) = 2285 MB/s
test bench::easy1_1MB                        ... bench:     433,971 ns/iter (+/- 262,128) = 2415 MB/s
test bench::easy1_32                         ... bench:         109 ns/iter (+/- 25) = 293 MB/s
test bench::easy1_32K                        ... bench:      12,397 ns/iter (+/- 1,170) = 2643 MB/s
test bench::hard_1K                          ... bench:      48,867 ns/iter (+/- 2,697) = 20 MB/s
test bench::hard_1MB                         ... bench: 111,206,218 ns/iter (+/- 5,323,278) = 8 MB/s
test bench::hard_32                          ... bench:       1,644 ns/iter (+/- 122) = 19 MB/s
test bench::hard_32K                         ... bench:   1,587,993 ns/iter (+/- 75,282) = 20 MB/s
test bench::literal                          ... bench:          25 ns/iter (+/- 2)
test bench::match_class                      ... bench:         133 ns/iter (+/- 7)
test bench::match_class_in_range             ... bench:         134 ns/iter (+/- 6)
test bench::match_class_unicode              ... bench:       3,300 ns/iter (+/- 234)
test bench::medium_1K                        ... bench:       3,368 ns/iter (+/- 915) = 304 MB/s
test bench::medium_1MB                       ... bench:   3,393,398 ns/iter (+/- 1,335,491) = 308 MB/s
test bench::medium_32                        ... bench:         193 ns/iter (+/- 173) = 165 MB/s
test bench::medium_32K                       ... bench:     107,107 ns/iter (+/- 58,285) = 305 MB/s
test bench::no_exponential                   ... bench:     475,996 ns/iter (+/- 352,366)
test bench::not_literal                      ... bench:       1,755 ns/iter (+/- 1,136)
test bench::one_pass_long_prefix             ... bench:         429 ns/iter (+/- 29)
test bench::one_pass_long_prefix_not         ... bench:         437 ns/iter (+/- 220)
test bench::one_pass_short_a                 ... bench:       1,744 ns/iter (+/- 1,922)
test bench::one_pass_short_a_not             ... bench:       1,564 ns/iter (+/- 495)
test bench::one_pass_short_b                 ... bench:       1,369 ns/iter (+/- 645)
test bench::one_pass_short_b_not             ... bench:       1,354 ns/iter (+/- 411)
test bench::replace_all                      ... bench:         249 ns/iter (+/- 12)

```