@bluss 2015-08-03
I have mixed feelings about this now. It's really a tradeoff — there are some arguably simple cases that are much worse off without this.

I went and looked some super simple ascii string searches (for my own code) and they all degraded with this change, now that I evaluated it for regex too. (While the original issue remains, that repeated memchr is 10x worse for certain periodic patterns).

[benchmark definitions, exact commit](https://github.com/bluss/scratchspace/blob/b61232f4fa7ec5eeca6a4e309ea052394bddfcf7/benches/pathology.rs#L254-L315)

With repeated memchr:

```
test aaab_in_aab::regex_find                ... bench:   4,148,943 ns/iter (+/- 9,933) = 72 MB/s
test aaabbb::regex_find                     ... bench:   4,099,691 ns/iter (+/- 23,054) = 72 MB/s
test periodic2::regex_find                  ... bench:  23,224,198 ns/iter (+/- 345,872)
test periodic5::regex_find                  ... bench:   4,206,946 ns/iter (+/- 136,301) = 1 MB/s
test short_1let_long::regex_find            ... bench:         161 ns/iter (+/- 1) = 15844 MB/s
test short_2let_long::regex_find            ... bench:         628 ns/iter (+/- 142) = 4062 MB/s
test short_3let_long::regex_find            ... bench:       3,602 ns/iter (+/- 421) = 708 MB/s
test short_short::regex_find                ... bench:         111 ns/iter (+/- 1) = 504 MB/s
test short_word1_long::regex_find           ... bench:       3,485 ns/iter (+/- 403) = 731 MB/s
test short_word2_long::regex_find           ... bench:       2,340 ns/iter (+/- 2) = 1090 MB/s
```

Without repeated memchr:

```
test aaab_in_aab::regex_find                ... bench:   1,246,280 ns/iter (+/- 48,122) = 240 MB/s
test aaabbb::regex_find                     ... bench:     446,015 ns/iter (+/- 6,960) = 672 MB/s
test periodic2::regex_find                  ... bench:  23,542,007 ns/iter (+/- 348,500)
test periodic5::regex_find                  ... bench:   4,316,395 ns/iter (+/- 116,745) = 1 MB/s
test short_1let_long::regex_find            ... bench:         161 ns/iter (+/- 2) = 15844 MB/s
test short_2let_long::regex_find            ... bench:      10,486 ns/iter (+/- 32) = 243 MB/s
test short_3let_long::regex_find            ... bench:       7,566 ns/iter (+/- 43) = 337 MB/s
test short_short::regex_find                ... bench:         120 ns/iter (+/- 1) = 466 MB/s
test short_word1_long::regex_find           ... bench:       3,697 ns/iter (+/- 13) = 690 MB/s
test short_word2_long::regex_find           ... bench:       6,604 ns/iter (+/- 50) = 386 MB/s
```


So this is not good, the "natural language" (maybe one should have a bigger better corpus for this?) cases were all made worse, while the silly pathologies were improved. We didn't catch this before because neither you nor regex had benchmark cases for it.

So one way to solve it is this an ugly contraption, which does in fact allow us to have it both ways. It works — it's just that it's definitely icky!

```diff
diff --git a/src/prefix.rs b/src/prefix.rs
index a738976..a8eb58b 100644
--- a/src/prefix.rs
+++ b/src/prefix.rs
@@ -187,6 +187,7 @@ impl SingleSearch {
     fn find(&self, haystack: &str) -> Option<usize> {
         let pat = self.pat.as_bytes();
         let haystack = haystack.as_bytes();
+        let mut use_memchr = 100;
         if haystack.len() < pat.len() {
             return None;
         }
@@ -203,6 +204,14 @@ impl SingleSearch {
                 return Some(i);
             }
             i += self.shift[b as usize];
+            if use_memchr > 0 {
+                let mc = match memchr(pat[0], &haystack[i..]) {
+                    None => return None,
+                    Some(i) => i,
+                };
+                use_memchr += if mc != 0 { 0 } else { -1 };
+                i += mc;
+            }
         }
         None
     }
```

I don't know what's best to do. I think this PR should be reverted, at least.